@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
 // 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.12
+// 	protoc        v5.29.3
 // source: proto/eupalinos.proto
 
 package proto
@@ -485,6 +485,108 @@ func (x *EnqueueResponse) GetId() string {
 	return ""
 }
 
+type HealthRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+}
+
+func (x *HealthRequest) Reset() {
+	*x = HealthRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_eupalinos_proto_msgTypes[8]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *HealthRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*HealthRequest) ProtoMessage() {}
+
+func (x *HealthRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_eupalinos_proto_msgTypes[8]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use HealthRequest.ProtoReflect.Descriptor instead.
+func (*HealthRequest) Descriptor() ([]byte, []int) {
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{8}
+}
+
+// Response message for health check
+type HealthResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Healthy bool   `protobuf:"varint,1,opt,name=healthy,proto3" json:"healthy,omitempty"`
+	Time    string `protobuf:"bytes,2,opt,name=time,proto3" json:"time,omitempty"`
+	Version string `protobuf:"bytes,3,opt,name=version,proto3" json:"version,omitempty"`
+}
+
+func (x *HealthResponse) Reset() {
+	*x = HealthResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_eupalinos_proto_msgTypes[9]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *HealthResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*HealthResponse) ProtoMessage() {}
+
+func (x *HealthResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_eupalinos_proto_msgTypes[9]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use HealthResponse.ProtoReflect.Descriptor instead.
+func (*HealthResponse) Descriptor() ([]byte, []int) {
+	return file_proto_eupalinos_proto_rawDescGZIP(), []int{9}
+}
+
+func (x *HealthResponse) GetHealthy() bool {
+	if x != nil {
+		return x.Healthy
+	}
+	return false
+}
+
+func (x *HealthResponse) GetTime() string {
+	if x != nil {
+		return x.Time
+	}
+	return ""
+}
+
+func (x *HealthResponse) GetVersion() string {
+	if x != nil {
+		return x.Version
+	}
+	return ""
+}
+
 var File_proto_eupalinos_proto protoreflect.FileDescriptor
 
 var file_proto_eupalinos_proto_rawDesc = []byte{
@@ -518,29 +620,39 @@ var file_proto_eupalinos_proto_rawDesc = []byte{
 	0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06, 0x6c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x22, 0x21,
 	0x0a, 0x0f, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
 	0x65, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69,
-	0x64, 0x2a, 0x25, 0x0a, 0x09, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0b,
-	0x0a, 0x07, 0x45, 0x4e, 0x51, 0x55, 0x45, 0x55, 0x45, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x44,
-	0x45, 0x51, 0x55, 0x45, 0x55, 0x45, 0x10, 0x01, 0x32, 0x83, 0x02, 0x0a, 0x09, 0x45, 0x75, 0x70,
-	0x61, 0x6c, 0x69, 0x6e, 0x6f, 0x73, 0x12, 0x48, 0x0a, 0x12, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x51, 0x75, 0x65, 0x75, 0x65, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x73, 0x12, 0x14, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x55, 0x70, 0x64, 0x61,
-	0x74, 0x65, 0x1a, 0x18, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x49, 0x6e, 0x74, 0x65, 0x72,
-	0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x28, 0x01, 0x30, 0x01,
-	0x12, 0x3a, 0x0a, 0x0e, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x12, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x70, 0x69, 0x73, 0x74,
-	0x65, 0x6c, 0x6c, 0x6f, 0x1a, 0x16, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x6e, 0x71,
-	0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x36, 0x0a, 0x0e,
-	0x44, 0x65, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x12,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x43, 0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e,
-	0x66, 0x6f, 0x1a, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45, 0x70, 0x69, 0x73, 0x74,
-	0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x38, 0x0a, 0x0e, 0x47, 0x65, 0x74, 0x51, 0x75, 0x65, 0x75, 0x65,
-	0x4c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x12, 0x12, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x43,
-	0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x12, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x2e, 0x51, 0x75, 0x65, 0x75, 0x65, 0x4c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x42, 0x2b,
-	0x5a, 0x29, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x6f, 0x64, 0x79,
-	0x73, 0x73, 0x65, 0x69, 0x61, 0x2d, 0x67, 0x72, 0x65, 0x65, 0x6b, 0x2f, 0x65, 0x75, 0x70, 0x61,
-	0x6c, 0x69, 0x6e, 0x6f, 0x73, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x33,
+	0x64, 0x22, 0x0f, 0x0a, 0x0d, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x52, 0x65, 0x71, 0x75, 0x65,
+	0x73, 0x74, 0x22, 0x58, 0x0a, 0x0e, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x52, 0x65, 0x73, 0x70,
+	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x68, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x79, 0x18,
+	0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x07, 0x68, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x79, 0x12, 0x12,
+	0x0a, 0x04, 0x74, 0x69, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x74, 0x69,
+	0x6d, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x03, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x2a, 0x25, 0x0a, 0x09,
+	0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0b, 0x0a, 0x07, 0x45, 0x4e, 0x51,
+	0x55, 0x45, 0x55, 0x45, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x44, 0x45, 0x51, 0x55, 0x45, 0x55,
+	0x45, 0x10, 0x01, 0x32, 0xbc, 0x02, 0x0a, 0x09, 0x45, 0x75, 0x70, 0x61, 0x6c, 0x69, 0x6e, 0x6f,
+	0x73, 0x12, 0x37, 0x0a, 0x06, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x12, 0x14, 0x2e, 0x70, 0x72,
+	0x6f, 0x74, 0x6f, 0x2e, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x1a, 0x15, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68,
+	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x48, 0x0a, 0x12, 0x53, 0x74,
+	0x72, 0x65, 0x61, 0x6d, 0x51, 0x75, 0x65, 0x75, 0x65, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x73,
+	0x12, 0x14, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x1a, 0x18, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x49,
+	0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x45, 0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f,
+	0x28, 0x01, 0x30, 0x01, 0x12, 0x3a, 0x0a, 0x0e, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d,
+	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45,
+	0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x1a, 0x16, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
+	0x2e, 0x45, 0x6e, 0x71, 0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x12, 0x36, 0x0a, 0x0e, 0x44, 0x65, 0x71, 0x75, 0x65, 0x75, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61,
+	0x67, 0x65, 0x12, 0x12, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x43, 0x68, 0x61, 0x6e, 0x6e,
+	0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x10, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x45,
+	0x70, 0x69, 0x73, 0x74, 0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x38, 0x0a, 0x0e, 0x47, 0x65, 0x74, 0x51,
+	0x75, 0x65, 0x75, 0x65, 0x4c, 0x65, 0x6e, 0x67, 0x74, 0x68, 0x12, 0x12, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x2e, 0x43, 0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x1a, 0x12,
+	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x51, 0x75, 0x65, 0x75, 0x65, 0x4c, 0x65, 0x6e, 0x67,
+	0x74, 0x68, 0x42, 0x31, 0x5a, 0x2f, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
+	0x2f, 0x6f, 0x64, 0x79, 0x73, 0x73, 0x65, 0x69, 0x61, 0x2d, 0x67, 0x72, 0x65, 0x65, 0x6b, 0x2f,
+	0x61, 0x67, 0x6f, 0x72, 0x61, 0x2f, 0x65, 0x75, 0x70, 0x61, 0x6c, 0x69, 0x6e, 0x6f, 0x73, 0x2f,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -556,7 +668,7 @@ func file_proto_eupalinos_proto_rawDescGZIP() []byte {
 }
 
 var file_proto_eupalinos_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_proto_eupalinos_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
+var file_proto_eupalinos_proto_msgTypes = make([]protoimpl.MessageInfo, 10)
 var file_proto_eupalinos_proto_goTypes = []interface{}{
 	(Operation)(0),            // 0: proto.Operation
 	(*InternalEpistello)(nil), // 1: proto.InternalEpistello
@@ -567,23 +679,27 @@ var file_proto_eupalinos_proto_goTypes = []interface{}{
 	(*Empty)(nil),             // 6: proto.Empty
 	(*QueueLength)(nil),       // 7: proto.QueueLength
 	(*EnqueueResponse)(nil),   // 8: proto.EnqueueResponse
+	(*HealthRequest)(nil),     // 9: proto.HealthRequest
+	(*HealthResponse)(nil),    // 10: proto.HealthResponse
 }
 var file_proto_eupalinos_proto_depIdxs = []int32{
-	0, // 0: proto.MessageUpdate.operation:type_name -> proto.Operation
-	1, // 1: proto.MessageUpdate.message:type_name -> proto.InternalEpistello
-	3, // 2: proto.Eupalinos.StreamQueueUpdates:input_type -> proto.MessageUpdate
-	2, // 3: proto.Eupalinos.EnqueueMessage:input_type -> proto.Epistello
-	5, // 4: proto.Eupalinos.DequeueMessage:input_type -> proto.ChannelInfo
-	5, // 5: proto.Eupalinos.GetQueueLength:input_type -> proto.ChannelInfo
-	1, // 6: proto.Eupalinos.StreamQueueUpdates:output_type -> proto.InternalEpistello
-	8, // 7: proto.Eupalinos.EnqueueMessage:output_type -> proto.EnqueueResponse
-	2, // 8: proto.Eupalinos.DequeueMessage:output_type -> proto.Epistello
-	7, // 9: proto.Eupalinos.GetQueueLength:output_type -> proto.QueueLength
-	6, // [6:10] is the sub-list for method output_type
-	2, // [2:6] is the sub-list for method input_type
-	2, // [2:2] is the sub-list for extension type_name
-	2, // [2:2] is the sub-list for extension extendee
-	0, // [0:2] is the sub-list for field type_name
+	0,  // 0: proto.MessageUpdate.operation:type_name -> proto.Operation
+	1,  // 1: proto.MessageUpdate.message:type_name -> proto.InternalEpistello
+	9,  // 2: proto.Eupalinos.Health:input_type -> proto.HealthRequest
+	3,  // 3: proto.Eupalinos.StreamQueueUpdates:input_type -> proto.MessageUpdate
+	2,  // 4: proto.Eupalinos.EnqueueMessage:input_type -> proto.Epistello
+	5,  // 5: proto.Eupalinos.DequeueMessage:input_type -> proto.ChannelInfo
+	5,  // 6: proto.Eupalinos.GetQueueLength:input_type -> proto.ChannelInfo
+	10, // 7: proto.Eupalinos.Health:output_type -> proto.HealthResponse
+	1,  // 8: proto.Eupalinos.StreamQueueUpdates:output_type -> proto.InternalEpistello
+	8,  // 9: proto.Eupalinos.EnqueueMessage:output_type -> proto.EnqueueResponse
+	2,  // 10: proto.Eupalinos.DequeueMessage:output_type -> proto.Epistello
+	7,  // 11: proto.Eupalinos.GetQueueLength:output_type -> proto.QueueLength
+	7,  // [7:12] is the sub-list for method output_type
+	2,  // [2:7] is the sub-list for method input_type
+	2,  // [2:2] is the sub-list for extension type_name
+	2,  // [2:2] is the sub-list for extension extendee
+	0,  // [0:2] is the sub-list for field type_name
 }
 
 func init() { file_proto_eupalinos_proto_init() }
@@ -688,6 +804,30 @@ func file_proto_eupalinos_proto_init() {
 				return nil
 			}
 		}
+		file_proto_eupalinos_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*HealthRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_proto_eupalinos_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*HealthResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
@@ -695,7 +835,7 @@ func file_proto_eupalinos_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_proto_eupalinos_proto_rawDesc,
 			NumEnums:      1,
-			NumMessages:   8,
+			NumMessages:   10,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
