@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.21.12
+// - protoc             v5.29.3
 // source: proto/eupalinos.proto
 
 package proto
@@ -22,6 +22,7 @@ const _ = grpc.SupportPackageIsVersion7
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type EupalinosClient interface {
+	Health(ctx context.Context, in *HealthRequest, opts ...grpc.CallOption) (*HealthResponse, error)
 	// Bidirectional Streaming for task updates between Eupalinos pods
 	StreamQueueUpdates(ctx context.Context, opts ...grpc.CallOption) (Eupalinos_StreamQueueUpdatesClient, error)
 	// Unary RPC for epistello enqueueing
@@ -38,6 +39,15 @@ type eupalinosClient struct {
 
 func NewEupalinosClient(cc grpc.ClientConnInterface) EupalinosClient {
 	return &eupalinosClient{cc}
+}
+
+func (c *eupalinosClient) Health(ctx context.Context, in *HealthRequest, opts ...grpc.CallOption) (*HealthResponse, error) {
+	out := new(HealthResponse)
+	err := c.cc.Invoke(ctx, "/proto.Eupalinos/Health", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
 }
 
 func (c *eupalinosClient) StreamQueueUpdates(ctx context.Context, opts ...grpc.CallOption) (Eupalinos_StreamQueueUpdatesClient, error) {
@@ -102,6 +112,7 @@ func (c *eupalinosClient) GetQueueLength(ctx context.Context, in *ChannelInfo, o
 // All implementations must embed UnimplementedEupalinosServer
 // for forward compatibility
 type EupalinosServer interface {
+	Health(context.Context, *HealthRequest) (*HealthResponse, error)
 	// Bidirectional Streaming for task updates between Eupalinos pods
 	StreamQueueUpdates(Eupalinos_StreamQueueUpdatesServer) error
 	// Unary RPC for epistello enqueueing
@@ -117,6 +128,9 @@ type EupalinosServer interface {
 type UnimplementedEupalinosServer struct {
 }
 
+func (UnimplementedEupalinosServer) Health(context.Context, *HealthRequest) (*HealthResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method Health not implemented")
+}
 func (UnimplementedEupalinosServer) StreamQueueUpdates(Eupalinos_StreamQueueUpdatesServer) error {
 	return status.Errorf(codes.Unimplemented, "method StreamQueueUpdates not implemented")
 }
@@ -140,6 +154,24 @@ type UnsafeEupalinosServer interface {
 
 func RegisterEupalinosServer(s grpc.ServiceRegistrar, srv EupalinosServer) {
 	s.RegisterService(&Eupalinos_ServiceDesc, srv)
+}
+
+func _Eupalinos_Health_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(HealthRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(EupalinosServer).Health(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/proto.Eupalinos/Health",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(EupalinosServer).Health(ctx, req.(*HealthRequest))
+	}
+	return interceptor(ctx, in, info, handler)
 }
 
 func _Eupalinos_StreamQueueUpdates_Handler(srv interface{}, stream grpc.ServerStream) error {
@@ -229,6 +261,10 @@ var Eupalinos_ServiceDesc = grpc.ServiceDesc{
 	ServiceName: "proto.Eupalinos",
 	HandlerType: (*EupalinosServer)(nil),
 	Methods: []grpc.MethodDesc{
+		{
+			MethodName: "Health",
+			Handler:    _Eupalinos_Health_Handler,
+		},
 		{
 			MethodName: "EnqueueMessage",
 			Handler:    _Eupalinos_EnqueueMessage_Handler,
